% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba-diags.R
\name{as_jabba_tidy}
\alias{as_jabba_tidy}
\title{Coerce JABBA fit to a standard list
Minimal assumption: we can build chains, residuals, ts from the raw fit.}
\usage{
as_jabba_tidy(raw_fit, stock = "", scenario = "base")
}
\description{
Coerce JABBA fit to a standard list
Minimal assumption: we can build chains, residuals, ts from the raw fit.
}
