% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivors.R
\name{b0dyn,FLStock,FLSR-method}
\alias{b0dyn,FLStock,FLSR-method}
\title{Survivorship from total mortality}
\usage{
\S4method{b0dyn}{FLStock,FLSR}(object, sr, residuals = 1, ...)
}
\arguments{
\item{object}{\code{FLSR}.}

\item{...}{Additional arguments.}
}
\description{
Survivorship from total mortality
}
