% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesdata-utils.R
\name{benchmarksFn}
\alias{benchmarksFn}
\title{Internal Function for Benchmark Extraction}
\usage{
benchmarksFn(object)
}
\arguments{
\item{object}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for Benchmark Extraction
}
\keyword{internal}
