% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat-sel.R
\name{calcOverlap}
\alias{calcOverlap}
\title{Normalized overlap between two age-specific curves}
\usage{
calcOverlap(curve1, curve2)
}
\arguments{
\item{curve1, curve2}{Numeric vectors of the same length.}
}
\value{
Proportion in \eqn{[0, 1]}.
}
\description{
Normalized overlap between two age-specific curves
}
