% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{calcPriors}
\alias{calcPriors}
\title{Calculate priors for an `FLBRP` based on reference points and observations.}
\usage{
calcPriors(object, ...)
}
\arguments{
\item{object}{`An`FLBRP` object for which to calculate priors}

\item{...}{Additional arguments passed to methods}
}
\value{
A named vector of calculated prior values
}
\description{
Calculate priors for an `FLBRP` based on reference points and observations.
}
