% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{calculateBenchmarks}
\alias{calculateBenchmarks}
\title{Calculate Benchmark Reference Points}
\usage{
calculateBenchmarks(icesData)
}
\arguments{
\item{icesdata}{Processed ICES stock data}
}
\value{
Data frame with benchmark reference points
}
\description{
Calculates benchmark reference points for ICES stocks
}
\examples{
bm = calculateBenchmarks(icesData)
}
