% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{calculateMpStats}
\alias{calculateMpStats}
\title{Calculate Management Plan Statistics}
\usage{
calculateMpStats(timeSeriesData)
}
\arguments{
\item{timeSeriesData}{Time series data}
}
\value{
Data frame with management plan statistics
}
\description{
Calculates management plan statistics for time series data
}
\examples{
mpStats = calculateMpStats(timeSeriesData)
}
