% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{calculateTimeseries}
\alias{calculateTimeseries}
\title{Calculate Time Series Data}
\usage{
calculateTimeseries(icesData, benchmarkData)
}
\arguments{
\item{icesData}{Processed ICES stock data}

\item{benchmarkData}{Benchmark reference points}
}
\value{
Merged time series and benchmark data
}
\description{
Calculates time series data for ICES stocks
}
\examples{
tsData = calculateTimeseries(icesData, benchmarkData)
}
