% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{checkVariation}
\alias{checkVariation}
\alias{checkVariation,FLStock-method}
\alias{checkVariation,FLStocks-method}
\title{Check Parameter Variation}
\usage{
checkVariation(object, ...)

\S4method{checkVariation}{FLStock}(object)

\S4method{checkVariation}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments}
}
\value{
A data.frame with logical values indicating variation:
  \item{mAge}{Logical: M varies by age}
  \item{mYr}{Logical: M varies by year}
  \item{massYr}{Logical: Stock weight varies by year}
  \item{matYr}{Logical: Maturity varies by year}
  For FLStocks, also includes .id column

A data.frame with logical values indicating variation:
  \item{mAge}{Logical: M varies by age}
  \item{mYr}{Logical: M varies by year}
  \item{massYr}{Logical: Stock weight varies by year}
  \item{matYr}{Logical: Maturity varies by year}
  For FLStocks, also includes .id column
}
\description{
Determines if biological parameters vary by age and/or year in an FLStock object.
Checks for variation in natural mortality (M), stock weight, and maturity.

Determines if biological parameters vary by age and/or year in an FLStock object.
Checks for variation in natural mortality (M), stock weight, and maturity.
}
\examples{
\dontrun{
data(icesdata)
vary = checkVariation(icesdata[[1]])
varyAll = checkVariation(icesdata)
}
}
