% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{crosstest}
\alias{crosstest}
\title{Cross-Test Analysis}
\usage{
crosstest(object, ...)
}
\arguments{
\item{object}{An object to analyze}

\item{...}{Additional arguments}
}
\description{
Generic function for cross-test analysis
}
