% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctc1903}
\alias{ctc1903}
\title{ICES Historical Catch Data (1903 onwards)}
\format{
A data frame with 9292 observations and 3 variables:
  \itemize{
    \item .id - Stock identifier (character)
    \item year - Year (integer, starting from 1903)
    \item catch - Catch values (integer)
  }
}
\source{
ICES Working Group on Stock Assessment Methods
}
\usage{
data(ctc1903)
}
\description{
Historical catch data for ICES stocks starting from 1903.
Contains catch time series for multiple stocks.
}
\examples{
\dontrun{
data(ctc1903)
head(ctc1903)
summary(ctc1903)
}
}
\keyword{datasets}
