% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{ebiomass}
\alias{ebiomass}
\alias{ebiomass,FLStock-method}
\alias{ebiomass,FLBRP-method}
\title{Calculate Exploitable Biomass}
\usage{
ebiomass(object)

\S4method{ebiomass}{FLStock}(object)

\S4method{ebiomass}{FLBRP}(object)
}
\arguments{
\item{object}{An object of class FLStock or FLBRP}
}
\value{
An FLQuant object containing the exploitable biomass time series

An FLQuant object containing the exploitable biomass time series
}
\description{
Calculates the exploitable biomass from an FLStock object using selectivity-weighted
catch weights and stock numbers.

Calculates the exploitable biomass from FLStock or FLBRP objects using selectivity-weighted
catch weights and stock numbers.
}
\details{
Exploitable biomass is based on weighting catch weights by selectivity normalised 
by peak selectivity, then multiplying by stock numbers and summing across ages

Exploitable biomass is based on weighting catch weights by selectivity normalised 
by peak selectivity, then multiplying by stock numbers and summing across ages
}
\examples{
\dontrun{
data(ple4)
eb <- ebiomass(ple4)
}

\dontrun{
data(ple4)
eb <- ebiomass(ple4)
}

}
