% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{fishlife}
\alias{fishlife}
\alias{fishlife,FLStock-method}
\alias{fishlife,FLStocks-method}
\title{Calculate something using the haupt method for FLQuant and FLPar objects.}
\usage{
fishlife(object, ...)

\S4method{fishlife}{FLStock}(object)

\S4method{fishlife}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}

\item{pars}{A parameter object (class FLPar) containing necessary parameters.}

\item{lc}{A threshold value for lc.}

\item{lmax}{A threshold value for lmax.}
}
\value{
The result of the calculation.

An FLPar object containing FishLife parameters
}
\description{
A generic function to extract FishLife parameters from FLStock objects
}
\details{
This method calculates something using the `haupt` function when the input object is of class FLQuant and the parameter object is of class FLPar.
}
