% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesdata-utils.R
\name{fishlifesFn}
\alias{fishlifesFn}
\title{Internal Function for FishLife Parameter Extraction}
\usage{
fishlifesFn(object)
}
\arguments{
\item{object}{An FLStock object}
}
\value{
An FLPar object
}
\description{
Internal Function for FishLife Parameter Extraction
}
\keyword{internal}
