% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{forage}
\alias{forage}
\title{Forage Index}
\usage{
forage(object, ...)
}
\arguments{
\item{object}{An object to analyze}

\item{...}{Additional arguments}
}
\description{
Generic function for forage fish index calculation
}
