% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-functions.R
\name{fwdFmsy}
\alias{fwdFmsy}
\title{Calculate Forward FMSY Projections}
\usage{
fwdFmsy(stock, years, fmsy)
}
\arguments{
\item{stock}{FLStock object}

\item{years}{Years to project}

\item{fmsy}{FMSY value to use}
}
\value{
Projected stock
}
\description{
Calculates forward projections using FMSY reference points
}
