% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-helpers.R
\name{get_vignette_output_dir}
\alias{get_vignette_output_dir}
\title{Get Vignette Output Directory}
\usage{
get_vignette_output_dir()
}
\value{
Character vector with output directory path
}
\description{
Returns the path to the vignette output directory (data/vignette-outputs).
Creates the directory if it doesn't exist. This is a simplified version
specifically for saving vignette outputs.
}
\examples{
\dontrun{
vignetteDataDir <- get_vignette_output_dir()
save(myData, file = file.path(vignetteDataDir, "myData.RData"))
}
}
