% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{getNephropsCatchMultipliers}
\alias{getNephropsCatchMultipliers}
\title{Create Catch Multiplier Vector}
\usage{
getNephropsCatchMultipliers()
}
\value{
Vector of catch multipliers
}
\description{
Create Catch Multiplier Vector
}
