% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba-diags.R
\name{gg_external_ices}
\alias{gg_external_ices}
\title{JABBA B vs ICES SSB
ices_ts: year, SSB (or TB)}
\usage{
gg_external_ices(fit, ices_ts)
}
\description{
JABBA B vs ICES SSB
ices_ts: year, SSB (or TB)
}
