% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba-diags.R
\name{gg_mcmc_trace_density}
\alias{gg_mcmc_trace_density}
\title{Trace + density plots}
\usage{
gg_mcmc_trace_density(fit, params = c("r", "K", "sigma.proc", "sigma.obs"))
}
\description{
Trace + density plots
}
