% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jabba-diags.R
\name{gg_prior_posterior}
\alias{gg_prior_posterior}
\title{Prior–posterior overlay using a supplied prior density
priors: named list of density functions, e.g. list(r = function(x) dlnorm(x, m, s), ...)}
\usage{
gg_prior_posterior(fit, params, priors, n_grid = 200)
}
\description{
Prior–posterior overlay using a supplied prior density
priors: named list of density functions, e.g. list(r = function(x) dlnorm(x, m, s), ...)
}
