% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{haupt}
\alias{haupt}
\title{Calculate total mortality-at-age (Z) using length data.}
\usage{
haupt(object, pars, lc, lmax, ...)
}
\arguments{
\item{object}{An object of class FLQuant or data.frame (depends on the method).}

\item{pars}{A parameter object (class FLPar) containing necessary parameters.}

\item{lc}{A threshold value for lc.}

\item{lmax}{A threshold value for lmax.}

\item{...}{Additional arguments .}
}
\value{
estimates of Z.
}
\description{
This is a generic S4 method that calculates something using the `haupt` function.
The specific implementation depends on the class of the input object.
}
