% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesdata}
\alias{icesdata}
\title{ICES Stock Assessment Data}
\format{
An \code{FLStocks} object containing multiple \code{FLStock} objects.
  Each stock contains:
  \itemize{
    \item Stock numbers at age (\code{stock.n})
    \item Catch numbers at age (\code{catch.n})
    \item Fishing mortality (\code{harvest})
    \item Natural mortality (\code{m})
    \item Stock weights (\code{stock.wt})
    \item Catch weights (\code{catch.wt})
    \item Maturity ogive (\code{mat})
    \item Selectivity (\code{harvest})
    \item Spawning stock biomass (\code{ssb})
    \item Catch (\code{catch})
  }
}
\source{
ICES Working Group on Stock Assessment Methods
}
\usage{
data(icesdata)
}
\description{
A dataset containing ICES Category 1 stock assessments (FLStocks object)
with analytical age-based assessments that have been benchmarked.
}
\examples{
\dontrun{
data(icesdata)
class(icesdata)  # Should be FLStocks
names(icesdata)  # List of stock names
}
}
\keyword{datasets}
