% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info}
\alias{info}
\title{ICES Stock Metadata}
\format{
A data frame with stock information including:
  \itemize{
    \item .id - Stock identifier
    \item Model - Assessment model type (e.g., "SMS", "XSA")
    \item Additional metadata columns
  }
}
\source{
ICES Working Group on Stock Assessment Methods
}
\usage{
data(info)
}
\description{
Metadata information for ICES stocks including stock identification,
assessment models, and other stock characteristics.
}
\examples{
\dontrun{
data(info)
head(info)
}
}
\keyword{datasets}
