% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{invALK}
\alias{invALK}
\title{Inverse Age-Length Key}
\usage{
invALK(
  object,
  model = NULL,
  age,
  cv = 0.1,
  lmax = 1.2,
  bin = 1,
  max = ceiling(object["linf"] * lmax),
  reflen = NULL,
  ...
)
}
\arguments{
\item{object}{An object containing length data}

\item{model}{Growth model function (default: NULL, will use vonbert from FLife if available)}

\item{age}{Age vector}

\item{cv}{Coefficient of variation (default: 0.1)}

\item{lmax}{Maximum length multiplier (default: 1.2)}

\item{bin}{Length bin size (default: 1)}

\item{max}{Maximum age (default: ceiling(object["linf"]*lmax))}

\item{reflen}{Reference length (optional)}

\item{...}{Additional arguments}
}
\description{
Generic function for inverse age-length key analysis
}
