% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenSamples.R
\docType{methods}
\name{lenSamples}
\alias{lenSamples}
\alias{lenSamples-method}
\alias{lenSamples,FLStock-method}
\title{lenSamples}
\usage{
lenSamples(object, par, n, ...)
}
\arguments{
\item{object}{\code{FLStock} object with catch data}

\item{par}{\code{FLPar} object with life history parameters (linf, k, l50)}

\item{n}{Numeric. Sample size for length frequencies}

\item{...}{any additional arguments}

\item{bin}{Numeric. Length bin width (default: 0.5)}

\item{extendToLinf}{Logical. Extend length bins to Linf * 1.1 (default: TRUE)}
}
\value{
\code{FLQuant} with length frequency data
}
\description{
Generates length frequency samples from an FLStock operating model.
Converts catch-at-age to length frequencies using inverse age-length keys.
}
\details{
Generate Length Samples from Operating Model
}
\examples{
\dontrun{
lsmps = lenSamples(om, par, n = 5000)
}
}
\author{
Based on code from generic.Rmd and case.Rmd
}
