% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{leslie}
\alias{leslie}
\title{Get priors for biomass dynamic assessment from an `FLBRP` object}
\usage{
leslie(object, fec, ...)
}
\arguments{
\item{object}{An object (typically an FLStock or FLBRP)}

\item{fec}{Fecundity parameters}

\item{...}{Additional arguments}
}
\value{
priors attribute as an `FLPar` of the `FLBRP` object
}
\description{
Generic function for Leslie matrix analysis
}
\examples{
\dontrun{
  # Create or load an FLBRP object
  # eql <- FLBRP(...)
  # priors(eql)
}
Leslie Matrix Analysis

}
