% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette-helpers.R
\name{load_ices_data}
\alias{load_ices_data}
\title{Load ICES Data from Package}
\usage{
load_ices_data(filename, subdir = NULL)
}
\arguments{
\item{filename}{Name of the RData file to load (e.g., "icesdata.RData")}

\item{subdir}{Optional subdirectory within vignettes/data}
}
\value{
Logical indicating success
}
\description{
Helper function to load ICES data files from the package or vignette directories.
This is primarily intended for use in vignettes.
}
\examples{
\dontrun{
# In a vignette
load_ices_data("icesdata.RData")
}
}
