% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{loadIcesRdata}
\alias{loadIcesRdata}
\title{Load ICES Data from RData File}
\usage{
loadIcesRdata(filePath = "data/icesdata.RData")
}
\arguments{
\item{filePath}{Path to ICES RData file}
}
\value{
Loaded ICES data
}
\description{
Load ICES Data from RData File
}
