% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{loadIcesStockData}
\alias{loadIcesStockData}
\title{Load ICES Stock Assessment Data}
\usage{
loadIcesStockData(
  filePath,
  speciesName = "Nephrops norvegicus",
  stockIndices = 3:12
)
}
\arguments{
\item{filePath}{Path to ICES data file}

\item{speciesName}{Species name to filter (default: "Nephrops norvegicus")}

\item{stockIndices}{Indices of stocks to select (default: 3:12)}
}
\value{
Processed stock data
}
\description{
Load ICES Stock Assessment Data
}
