% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loading.R
\name{loadStockCategories}
\alias{loadStockCategories}
\title{Load Stock Categories from Excel}
\usage{
loadStockCategories(filePath = "data/stocks.xls", sheetName = "cats")
}
\arguments{
\item{filePath}{Path to Excel file}

\item{sheetName}{Sheet name (default: "cats")}
}
\value{
Stock categories data
}
\description{
Load Stock Categories from Excel
}
