% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-helpers.R
\name{lor2gis}
\alias{lor2gis}
\title{Transform log(M)~log(W) linear relation to length-inverse mortality params}
\usage{
lor2gis(par)
}
\arguments{
\item{par}{An FLPar with at least params m1, linf, k}
}
\value{
An FLPar with elements m1, m2, m3, m4
}
\description{
Uses constants c=1.44 and d=1 to convert the linear relation parameters,
together with VB parameters linf and k, into the parameterization used
by inverse-length mortality models.
}
