% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-helpers.R
\name{lorFn}
\alias{lorFn}
\title{Log-linear mortality ~ weight relationship}
\usage{
lorFn(stk)
}
\arguments{
\item{stk}{An FLStock object}
}
\value{
Named numeric vector with elements m1 (intercept) and m2 (slope)
}
\description{
Regresses log(M) on log(weight) across ages to return intercept and slope.
}
