% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lw}
\alias{lw}
\title{ICES Length-Weight Relationships}
\format{
A data frame with species length-weight relationships:
  \itemize{
    \item name - Common name of species
    \item latin - Latin/scientific name
    \item a - Length-weight parameter a
    \item b - Length-weight parameter b
    \item Source - Data source (e.g., "FishBase", "ResearchGate")
  }
}
\source{
Various sources including FishBase and ResearchGate
}
\usage{
data(lw)
}
\description{
Length-weight relationship parameters for ICES fish species.
Parameters a and b in the equation W = a * L^b where W is weight and L is length.
}
\examples{
\dontrun{
data(lw)
head(lw)
}
}
\keyword{datasets}
