% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{m1}
\alias{m1}
\title{Natural Mortality M1}
\usage{
m1(object, ...)
}
\arguments{
\item{object}{An object to analyze}

\item{...}{Additional arguments}
}
\description{
Generic function for M1 natural mortality estimation
}
