% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{m2}
\alias{m2}
\title{Natural Mortality M2}
\usage{
m2(object, ...)
}
\arguments{
\item{object}{An object to analyze}

\item{...}{Additional arguments}
}
\description{
Generic function for M2 natural mortality (predation mortality) estimation
}
