% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{msyVirgin}
\alias{msyVirgin}
\title{Calculate MSY and Virgin State Metrics}
\usage{
msyVirgin(object, ...)
}
\arguments{
\item{object}{An object of class FLBRP}
}
\value{
A named vector containing metrics for virgin and MSY states:
  \item{virgin.f}{F at virgin state}
  \item{virgin.catch}{Catch at virgin state}
  \item{virgin.ebiomass}{Equilibrium biomass at virgin state}
  \item{msy.f}{F at MSY}
  \item{msy.catch}{Catch at MSY}
  \item{msy.ebiomass}{Equilibrium biomass at MSY}
}
\description{
Calculates key metrics (F, SSB, catch, ebiomass) at virgin and MSY states for an FLBRP object
}
\note{
The function excludes virgin.ssb and msy.ssb from the output
}
\examples{
\dontrun{
data(ple4brp)
msyVirgin(ple4brp)
}
}
