% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotBenchmarks}
\alias{plotBenchmarks}
\title{Plot Benchmark Reference Points}
\usage{
plotBenchmarks(benchmarkData)
}
\arguments{
\item{benchmarkData}{Benchmark data}
}
\value{
ggplot object
}
\description{
Creates a plot showing benchmark reference points
}
