% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotFishingMortality}
\alias{plotFishingMortality}
\title{Plot Fishing Mortality}
\usage{
plotFishingMortality(data, x = "year", y = "f", group = ".id")
}
\arguments{
\item{data}{Stock data}

\item{x}{Column name for x-axis (default: "year")}

\item{y}{Column name for y-axis (default: "f")}

\item{group}{Column name for grouping (default: ".id")}
}
\value{
ggplot object
}
\description{
Creates a plot showing fishing mortality over time
}
