% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plotLengths}
\alias{plotLengths}
\title{Plot Length Frequency Data}
\usage{
plotLengths(x, ...)
}
\arguments{
\item{x}{An object to plot}

\item{...}{Additional arguments}
}
\description{
Plot Length Frequency Data
}
