% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotManagementPlanStats}
\alias{plotManagementPlanStats}
\title{Plot Management Plan Statistics}
\usage{
plotManagementPlanStats(managementPlanStats)
}
\arguments{
\item{managementPlanStats}{Management plan statistics data}
}
\value{
ggplot object
}
\description{
Creates a stacked area plot showing management plan statistics
}
