% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat-sel.R
\name{plotMatSel}
\alias{plotMatSel}
\title{Scatterplot of age at 50\% selectivity vs maturity}
\usage{
plotMatSel(x)
}
\arguments{
\item{x}{Data frame with columns \code{mat50} and \code{sel50} (e.g. from
\code{\link{selMetrics}}).}
}
\value{
A \code{ggplot} object.
}
\description{
Scatterplot of age at 50\% selectivity vs maturity
}
