% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{PN}
\alias{PN}
\title{Confusion Matrix Statistics}
\usage{
PN(object, y)
}
\arguments{
\item{y}{Numeric vector of actual values}

\item{x}{Numeric vector of predicted values}
}
\value{
A data frame containing TP, TN, FP, and FN counts
}
\description{
Calculates True Positive (TP), True Negative (TN), False Positive (FP), 
and False Negative (FN) counts from predicted and actual values
}
\examples{
\dontrun{
pred <- c(1, -1, 1, -1)
actual <- c(1, -1, -1, 1)
confusion_stats <- PN(pred, actual)
}

}
