% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{predNeed}
\alias{predNeed}
\title{Predator Need}
\usage{
predNeed(object, ...)
}
\arguments{
\item{object}{An object to analyze}

\item{...}{Additional arguments}
}
\description{
Generic function for predator need calculation
}
