% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{priors}
\alias{priors}
\title{priors}
\usage{
priors(object, ...)
}
\arguments{
\item{object}{An `FLBRP` from which to extract priors}

\item{...}{additional arguments passed to methods}
}
\value{
`priors` attribute as an `FLPar`
}
\description{
priors for biomass dynamic assessment from an `FLBRP` object
}
