% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{processIcesdata}
\alias{processIcesdata}
\title{Process ICES Stock Data (Backwards Compatibility)}
\usage{
processIcesdata(dataPath, infoPath)
}
\arguments{
\item{dataPath}{Path to the ICES data file}

\item{infoPath}{Path to the info file}
}
\value{
Processed ICES stock data
}
\description{
Loads and processes ICES stock data with standard cleaning operations.
This function provides backwards compatibility with the original name.
For new code, use \code{processIcesData} (with capital D) from data-loading.R
}
\examples{
Icesdata = processIcesdata("path/to/data.RData", "path/to/info.RData")
}
