% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{pt}
\alias{pt}
\alias{pt-method}
\alias{pt,FLPar-method}
\alias{pt,numeric-method}
\alias{pt,FLBRP-method}
\title{Production Type Method for Surplus Production Models}
\usage{
pt(object, ...)
}
\arguments{
\item{object}{Input object (FLPar, numeric, or FLBRP)}

\item{...}{Additional arguments passed to methods}

\item{shape}{Ratio of BMSY to virgin biomass (default: object["bmsy"]/object["v"])}

\item{what}{Function to calculate biomass metric (default: ssb)}
}
\value{
An FLPar object containing:
- r: intrinsic growth rate
- k: carrying capacity
- p: shape parameter
- msy: maximum sustainable yield
- bmsy: biomass at MSY
- fmsy: fishing mortality at MSY
- virgin: unfished biomass
}
\description{
Estimates parameters of surplus production models using the PT method by deriving
parameters from reference points and biomass ratios
}
\details{
Production Type (PT) Method
}
\examples{
\dontrun{
params <- FLPar(c(bmsy=500, msy=125, v=1000))
pt(params)
}

}
