% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp}
\alias{spp}
\title{ICES Species Information}
\format{
A data frame with species information including:
  \itemize{
    \item Species identifiers
    \item Taxonomic information
    \item Biological characteristics
  }
}
\source{
ICES Working Group on Stock Assessment Methods
}
\usage{
data(spp)
}
\description{
Species information for ICES stocks including taxonomic and biological
characteristics.
}
\examples{
\dontrun{
data(spp)
head(spp)
}
}
\keyword{datasets}
