% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/tseries.R
\name{stdz}
\alias{stdz}
\alias{stdz,numeric-method}
\alias{stdz,matrix-method}
\alias{stdz,array-method}
\alias{stdz,FLQuant-method}
\title{Standardize Values}
\usage{
stdz(object, na.rm = TRUE, ...)

\S4method{stdz}{numeric}(object, na.rm = TRUE)

\S4method{stdz}{matrix}(object, na.rm = TRUE)

\S4method{stdz}{array}(object, na.rm = TRUE)

\S4method{stdz}{FLQuant}(object, na.rm = TRUE)
}
\arguments{
\item{object}{An object of class FLStock}

\item{na.rm}{Logical indicating whether to remove NA values when computing statistics}

\item{x}{A numeric vector, matrix, array or FLQuant}
}
\value{
An object of the same class as the input with standardized values

An FLQuant object containing the exploitable biomass time series
}
\description{
Standardizes values by subtracting the mean and dividing by the standard deviation.

Calculates the exploitable biomass from an FLStock object using selectivity-weighted
catch weights and stock numbers.
}
\details{
Standardization follows the formula: (x - mean(x))/sd(x)

Exploitable biomass is based on weighting catch weights by selectivity normalised 
by peak selectivity, then multiplying by stock numbers and summing across ages
}
\examples{
\dontrun{
# For numeric vector
x <- 1:10
stdz(object)

# For FLQuant
data(ple4)
standardized_catch <- stdz(catch(ple4))
}

\dontrun{
data(ple4)
eb <- ebiomass(ple4)
}
}
