% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivors.R
\name{survivors,FLQuant,missing-method}
\alias{survivors,FLQuant,missing-method}
\title{Survivorship from total mortality}
\usage{
\S4method{survivors}{FLQuant,missing}(object, sr, ...)
}
\arguments{
\item{object}{\code{FLQuant} of total mortality, \code{FLStock}, or \code{FLBRP}.}

\item{...}{Additional arguments.}

\item{cohort}{Logical; cohort-based survivorship for stock/BRP objects.}
}
\description{
Survivorship from total mortality
}
