% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSamIndex.R
\name{validateSamIndex}
\alias{validateSamIndex}
\title{Validate FLIndex structure for SAM}
\usage{
validateSamIndex(idx)
}
\arguments{
\item{idx}{FLIndex object to validate}
}
\value{
Logical - TRUE if valid, stops with error message if invalid
}
\description{
Validate FLIndex structure for SAM
}
