% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/icesdata-utils.R
\name{FLifePar}
\alias{FLifePar}
\alias{FLifePar,FLStock-method}
\alias{FLifePar,FLStocks-method}
\title{Internal Function for FishLife Parameter Extraction}
\usage{
FLifePar(object, ...)

\S4method{FLifePar}{FLStock}(object)

\S4method{FLifePar}{FLStocks}(object)
}
\arguments{
\item{object}{An FLStock or FLStocks object}

\item{...}{Additional arguments (not currently used)}

\item{x}{An FLStock object}
}
\value{
An FLPar object


Extract FLife Parameters

An FLPar object containing life history parameters
}
\description{
A generic function to extract FLife parameters from FLStock objects
}
